Move `obtain_current_passphrase` and `unlock_keyring` into `chia.cmds`
